Add Eslint and Prettier
