+++ src/containers/Pages/home.tsx
@@ -4,6 +4,12 @@ const Home: FunctionComponent = () => {
   return (
     <div>
       <h1>Home</h1>
+      <p>
+        Lorem ipsum dolor sit amet consectetur adipisicing elit. Omnis
+        perferendis ab sed dolor vitae. Neque enim dicta ea animi iste corrupti
+        et earum adipisci. Labore corrupti dicta excepturi laudantium
+        perferendis.
+      </p>
     </div>
   );
 };
